+++ burnman/__init__.py
@@ -205,6 +205,7 @@ from . import eos
 from . import processchemistry
 from . import composition_fitting
 from . import polytope
+from . import polytopetools
 from . import chemicalpotentials
 from . import geotherm
